Added a Github icon as an IconButton to link to project repo, as well as converting Social media Icons to IconButtons to allow functionality. FB, Insta, and Twitter icons currently have disabled class applied to make it clearer only the Github icon has a link. Moved Footer, FooterItems, and AdoptionForm to Components folder.
